services/producer-service/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
python-decouple>=3.8
PyJWT>=2.8.0
requests>=2.31.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0